fix(api): properly define indexes and move migrate
Refs: #57
